Add datetime to payment link title; fix amount sent as Int not Double
- Title now includes timestamp: "Appointment with Dr X - Y 2026-05-30 19-32"
- IntaSend requires amount as integer; convert Double->Int before API call
  (sending 750.0 caused "Problem experienced" 502 error)

+++ payments/src/main/kotlin/ke/co/smartroundclinic/payments/data/remote/dto/request/CreatePaymentLinkReq.kt
@@ -9,7 +9,7 @@ data class CreatePaymentLinkReq(
     val title: String,
     @SerialName("is_active") val isActive: Boolean? = null,
     @SerialName("redirect_url") val redirectUrl: String? = null,
-    val amount: Double? = null,
+    val amount: Int? = null,
     @SerialName("usage_limit") val usageLimit: Int? = null,
     val currency: String? = null,
     @SerialName("mobile_tarrif") val mobileTarrif: String? = null,
